REF: fixing Emitter type and detaching it from Action by introducing makeEmitter() fn.
