+++ example/index.ts
@@ -1,6 +1,7 @@
 import { createServer } from "../src";
 import { attachSockets } from "../src/sockets";
 import { config } from "./config";
+import { emission } from "./factories";
 import { actions, routing } from "./routing";
 import { Server } from "socket.io";
 
@@ -16,4 +17,5 @@ attachSockets({
   target: httpServer,
   actions,
   logger,
+  emission,
 });
